updateSecurityAttribute
