+++ lib/templates.ts
@@ -9,58 +9,6 @@ import { validateTemplate } from "@lib/utils/form-builder/validate";
 import { dateHasPast } from "@lib/utils";
 import { validateTemplateSize } from "@lib/utils/validateTemplateSize";
 
-export const updateSecurityAttribute = async (formID: string, securityAttribute: string) => {
-  const { user } = await authorization.canEditForm(formID).catch((e) => {
-    logEvent(
-      e.user.id,
-      { type: "Form", id: formID },
-      "AccessDenied",
-      AuditLogAccessDeniedDetails.AccessDenied_AttemptToUpdateSecurityAttribute
-    );
-    throw e;
-  });
-
-  const updatedTemplate = await prisma.template
-    .update({
-      where: {
-        id: formID,
-        isPublished: false,
-      },
-      data: { securityAttribute },
-      select: {
-        id: true,
-        created_at: true,
-        updated_at: true,
-        name: true,
-        jsonConfig: true,
-        isPublished: true,
-        deliveryOption: true,
-        securityAttribute: true,
-        formPurpose: true,
-        publishReason: true,
-        publishFormType: true,
-        publishDesc: true,
-        saveAndResume: true,
-        notificationsInterval: true,
-      },
-    })
-    .catch((e) => prismaErrors(e, null));
-
-  if (updatedTemplate === null) return updatedTemplate;
-
-  if (formCache.cacheAvailable) formCache.invalidate(formID);
-
-  logEvent(
-    user.id,
-    { type: "Form", id: formID },
-    AuditLogEvent.ChangeSecurityAttribute,
-    AuditLogDetails.ChangeSecurityAttribute,
-    { securityAttribute: securityAttribute ?? "" }
-  );
-
-  return _parseTemplate(updatedTemplate);
-};
-
 export const checkIfClosed = async (formId: string) => {
   try {
     let isPastClosingDate = false;
