+++ lib/templates/updateSecurityAttribute.ts
@@ -1,0 +1,63 @@
+import { formCache } from "../cache/formCache";
+import { prisma, prismaErrors } from "@gcforms/database";
+
+import { authorization } from "../privileges";
+import {
+  AuditLogAccessDeniedDetails,
+  AuditLogDetails,
+  AuditLogEvent,
+  logEvent,
+} from "../auditLogs";
+import { parseTemplate } from "./shared";
+
+export const updateSecurityAttribute = async (formID: string, securityAttribute: string) => {
+  const { user } = await authorization.canEditForm(formID).catch((e) => {
+    logEvent(
+      e.user.id,
+      { type: "Form", id: formID },
+      "AccessDenied",
+      AuditLogAccessDeniedDetails.AccessDenied_AttemptToUpdateSecurityAttribute
+    );
+    throw e;
+  });
+
+  const updatedTemplate = await prisma.template
+    .update({
+      where: {
+        id: formID,
+        isPublished: false,
+      },
+      data: { securityAttribute },
+      select: {
+        id: true,
+        created_at: true,
+        updated_at: true,
+        name: true,
+        jsonConfig: true,
+        isPublished: true,
+        deliveryOption: true,
+        securityAttribute: true,
+        formPurpose: true,
+        publishReason: true,
+        publishFormType: true,
+        publishDesc: true,
+        saveAndResume: true,
+        notificationsInterval: true,
+      },
+    })
+    .catch((e) => prismaErrors(e, null));
+
+  if (updatedTemplate === null) return updatedTemplate;
+
+  if (formCache.cacheAvailable) formCache.invalidate(formID);
+
+  logEvent(
+    user.id,
+    { type: "Form", id: formID },
+    AuditLogEvent.ChangeSecurityAttribute,
+    AuditLogDetails.ChangeSecurityAttribute,
+    { securityAttribute: securityAttribute ?? "" }
+  );
+
+  return parseTemplate(updatedTemplate);
+};
